fix examples; use async.forEachSeries
The examples were using async.forEach as a
way to asynchronously iterate over the files
to add them in the archives. This is problematic
because the library expects one file packaging
to be over before calling .addFile again.

Using forEachSeries fixes this, because it
iterates in series, not in parallel.