Add a `process.emitWarning` for `deprecated` (#202)
* Add a console.log call for @deprecated

The Archiver.bulk() call has been deprecated for a while and various
answers on SO and other sites still suggest it, so if there are no
planned maintenance fixes and updates to that function, everyone (even
those who do not read the docs) should be informed that they are using
something that is deprecated.

* Prefer process.emitWarning for @deprecated

The process.emitWarning call is much fancier than the plain logging
call as the user can --throw-deprecation or --no-deprecation if he
or she prefers to. As a fallback, we just console.warn to stderr.

* Inform about @deprecated only once

The deprecation logging action should only take place once so that
the client does not get spammed with the warnings. Additionally, the
console is now always expected to be around.

* Do not repeat the same string twice

Code maintainability and clean'ness and such features are included
in this commit.